issue=#319 Tera leveldb unit test

+++ src/leveldb/db/version_set_test.cc
@@ -21,7 +21,7 @@ class FindFileTest {
   FindFileTest() : disjoint_sorted_files_(true) { }
 
   ~FindFileTest() {
-    for (int i = 0; i < files_.size(); i++) {
+    for (size_t i = 0; i < files_.size(); i++) {
       delete files_[i];
     }
   }
